+++ backend/StudentManagement.Application/DTOs/Dashboard/DashboardDtos.cs
@@ -73,7 +73,7 @@ public class EnrollmentTrendDto
 
 public class StudentApplicationDto
 {
-    public int StudentId { get; set; }
+     public long StudentId { get; set; }
     public string StudentName { get; set; } = string.Empty;
     public int CourseId { get; set; }
     public string CourseName { get; set; } = string.Empty;
