+++ backend/StudentManagement.Application/Interfaces/IInstructorService.cs
@@ -6,7 +6,7 @@ namespace StudentManagement.Application.Interfaces;
 public interface IInstructorService
 {
     Task<PaginatedResponse<InstructorListItemDto>> GetInstructorsAsync(InstructorQueryParameters query);
-    Task<InstructorDto> GetInstructorAsync(int ssn, string? userId, IEnumerable<string> roles);
+    Task<InstructorDto> GetInstructorAsync(long ssn, string? userId, IEnumerable<string> roles);
     Task<InstructorDto> CreateInstructorAsync(CreateInstructorRequest request);
-    Task<InstructorDto> UpdateInstructorAsync(int ssn, UpdateInstructorRequest request);
+    Task<InstructorDto> UpdateInstructorAsync(long ssn, UpdateInstructorRequest request);
 }
