+++ backend/StudentManagement.Application/Interfaces/ITripService.cs
@@ -6,13 +6,13 @@ public interface ITripService
 {
     Task<TripDto?> GetByIdAsync(int tripId);
     Task<IEnumerable<TripDto>> GetAllAsync();
-    Task<IEnumerable<TripDto>> GetByDriverSsnAsync(int driverSsn);
-    Task<IEnumerable<TripDto>> GetByStudentSsnAsync(int studentSsn);
+    Task<IEnumerable<TripDto>> GetByDriverSsnAsync(long driverSsn);
+    Task<IEnumerable<TripDto>> GetByStudentSsnAsync(long studentSsn);
 
     Task<TripDto> CreateAsync(CreateTripDto dto);
     Task<TripDto> UpdateAsync(int tripId, UpdateTripDto dto);
     Task DeleteAsync(int tripId);
 
     Task<TripDto> AddStudentAsync(int tripId, AddStudentToTripDto dto);
-    Task RemoveStudentAsync(int tripId, int studentSsn);
+    Task RemoveStudentAsync(int tripId, long studentSsn);
 }
