+++ backend/StudentManagement.Domain/Entities/Instructor.cs
@@ -4,7 +4,7 @@ namespace StudentManagement.Domain.Entities;
 
 public class Instructor
 {
-    public int       InstructorSsn  { get; set; }
+    public long       InstructorSsn  { get; set; }
     public string    FirstName      { get; set; } = string.Empty;
     public string    LastName       { get; set; } = string.Empty;
     public string?   Phone          { get; set; }
